requirements38.txt
# # habitat-lab/requirements.txt
# gym==0.21.0  # gym>=0.17.3 (latest version not compatible)
numpy>=1.16.1
yacs>=0.1.5
numpy-quaternion>=2019.3.18.14.33.20
attrs>=19.1.0
opencv-python==4.5.5.64
pickle5; python_version < '3.8'
# visualization optional dependencies
imageio>=2.2.0
imageio-ffmpeg>=0.2.0
scipy>=1.0.0
tqdm>=4.0.0
numba>=0.44.0

# # habitat-lab/habitat_baselines/rl/requirements.txt
moviepy>=1.0.1
# torch>=1.3.1
# full tensorflow required for tensorboard video support
# tensorflow==1.13.1
# tb-nightly
tensorboardX>=2.0

# # habitat-lab/habitat_baselines/rl/ddppo/requirements.txt
ifcfg
# # VLN-CE
# attrs>=19.1.0
dtw==1.4.0
fastdtw==0.3.4
gdown
# gym>=0.17.3
jsonlines==2.0.0
lmdb
msgpack_numpy==0.4.7.1
networkx==2.5.1
numpy>=1.16.1
pre-commit
# torch>=1.6.0
# torchvision==0.2.2.post3
# tqdm>=4.0.0
# tensorflow==1.13.1
# tb-nightly
# yacs>=0.1.5
# Opencv-python>=3.3.0

# # Ours
boto3==1.18.17
transformers==4.9.2
pytorch-transformers
